basics work, somewhat

+++ network/events.go
@@ -1,7 +1,7 @@
 package network
 
 type Event struct {
-	senderID    int64
-	currentNode int64
-	nextNode    int64
+	senderID    int
+	currentNode int
+	nextNode    int
 }
